# swagger.yaml
openapi: 3.0.3
info:
    title: Swagger app Room
    description: api app Trọ trung chánh
    version: 1.0.11
servers:
    - url: https://api-app-room.onrender.com/
...
paths:
...
    /apiroom/updateguest/:
        put:
            tags:
                - room
            summary: Insert guest in room
            operationId: insertguest
            requestBody:
                description: Insert an existent guest in room
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Guest'
                    application/xml:
                        schema:
                            $ref: '#/components/schemas/Guest'
                    application/x-www-form-urlencoded:
                        schema:
                            $ref: '#/components/schemas/Guest'
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Guest'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Guest'
                default:
                    description: successful operation
    /apiroom/updateroom/:
        put:
            tags:
                - room
            summary: Update room
            operationId: updateroom
            requestBody:
                description: update an existent room
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateRoom'
                    application/xml:
                        schema:
                            $ref: '#/components/schemas/UpdateRoom'
                    application/x-www-form-urlencoded:
                        schema:
                            $ref: '#/components/schemas/UpdateRoom'
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateRoom'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/UpdateRoom'
                default:
                    description: successful operation
    /apiroom/create/:
        post:
            tags:
                - room
            summary: create new room
            operationId: createroom
            requestBody:
                content:
                    application/json:
                        schema:
                            type: array
                            items:
                                $ref: '#/components/schemas/Createroom'
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateroomSuccfully'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/CreateroomSuccfully'
                default:
                    description: successful operation
components:
    schemas:
...
        Guest:
            type: object
            properties:
                id: 
                    type: string
                    example: 6446702bc2c9dd70a6725e2d
                khachThue:
                    type: object
                    properties:
                        fullName:
                            type: string
                            example: "Nguyen Quang Huy"
                        phone:
                            type: string
                            example: "0364254852"
                        cccd:
                            type: string
                            example: "0015145224"
                        ngaySinh:
                            type: string
                            example: "19-06-2000"
                        diaChi:
                            type: string
                            example: "go vap, Ho Chi Minh"
...
        UpdateRoom:
            type: object
            properties:
                id: 
                    type: string
                    example: "6446702bc2c9dd70a6725e2d"
                name:
                    type: string
                    example: "Phòng 201"
                tienPhong:
                    type: number
                    example: 23564000
        Createroom: 
            type: object
            properties:
                khu: 
                    type: string
                    example: "a1dad1a"
                    required: true
                name: 
                    type: string
                    example: "Phòng 101"
                    required: true
                tienPhong: 
                    type: number
                    example: 2485642
                    required: true
        CreateroomSuccfully: 
            type: object
            properties:
                id: 
                    type: string
                    example: "a1dad1a"
                khu: 
                    type: string
                    example: "a1dad1a"
                name: 
                    type: string
                    example: "Phòng 101"
                tienPhong: 
                    type: number
                    example: 2485642
                tienThueThang:
                    type: array
                    example: []
                khachThue:
                    type: array
                    example: []
                initializationDate:
                    type: string
                    example: "2023-05-01"
                expirationDate:
                    type: string
                    example: "2023-06-01"
...
    securitySchemes:
        petstore_auth:
            type: oauth2
            flows:
                implicit:
                    authorizationUrl: https://petstore3.swagger.io/oauth/authorize
                    scopes:
                        write:pets: modify pets in your account
                        read:pets: read your pets
        api_key:
            type: apiKey
            name: api_key
            in: header